+++ Farmacia/src/br/com/thordrugstore/farmacia/model/Produto.java
@@ -45,7 +45,7 @@ public class Produto {
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
